pyproject.toml
[project]
name = "a2a-network"
version = "0.1.0"
description = "A2A protocol debate network - University project"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http,sqlite]==1.0.0a0",
    "litellm>=1.55",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sse-starlette>=2.0",
    "packaging",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"
